+++ grunt/config/gitcheckout.js
@@ -1,0 +1,29 @@
+/**
+ * Generated header by Cozen for generator-cozen-angular project
+ * Generated file gitcheckout on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 22/12/2017
+ * Time: 15:24
+ * Version: 1.0.0
+ */
+module.exports = {
+	master : {
+		options: {
+			verbose  : true,
+			branch   : 'master',
+			create   : false,
+			force    : false,
+			overwrite: false
+		}
+	},
+	develop: {
+		options: {
+			verbose  : true,
+			branch   : 'develop',
+			create   : false,
+			force    : false,
+			overwrite: false
+		}
+	}
+};