+++ grunt/config/gitmerge.js
@@ -1,0 +1,20 @@
+/**
+ * Generated header by Cozen for generator-cozen-angular project
+ * Generated file gitmerge on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 22/12/2017
+ * Time: 15:24
+ * Version: 1.0.0
+ */
+module.exports = {
+	develop: {
+		options: {
+			verbose : true,
+			branch  : 'develop',
+			message : 'Release <%= newVersion %>',
+			noCommit: false,
+			strategy: 'recursive'
+		}
+	}
+};