+++ grunt/config/gitpush.js
@@ -1,0 +1,33 @@
+/**
+ * Generated header by Cozen for generator-cozen-angular project
+ * Generated file gitpush on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 22/12/2017
+ * Time: 15:25
+ * Version: 1.0.0
+ */
+module.exports = {
+	master : {
+		options: {
+			verbose : true,
+			remote  : 'origin',
+			branch  : 'master',
+			all     : false,
+			tags    : true,
+			upstream: true,
+			force   : false
+		}
+	},
+	develop: {
+		options: {
+			verbose : true,
+			remote  : 'origin',
+			branch  : 'develop',
+			all     : false,
+			tags    : true,
+			upstream: true,
+			force   : false
+		}
+	}
+};