+++ grunt/config/gittag.js
@@ -1,0 +1,21 @@
+/**
+ * Generated header by Cozen for generator-cozen-angular project
+ * Generated file gittag on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 22/12/2017
+ * Time: 15:25
+ * Version: 1.0.0
+ */
+module.exports = {
+	release: {
+		options: {
+			verbose  : true,
+			tag      : '<%= newVersion %>',
+			message  : 'Version <%= newVersion %>',
+			remove   : false,
+			annotated: false,
+			force    : false
+		}
+	}
+};